requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-braces==1.4.0
django-compressor==1.4
djangorestframework==3.0.0
drf-nested-routers==0.9.0
gunicorn==19.1.1
oauthlib==0.7.2
six==1.8.0
static3==0.5.1
